+++ functions.php
@@ -181,10 +181,39 @@ add_filter( 'wp_resource_hints', 'cover2_resource_hints', 10, 2 );
  * @link https://developer.wordpress.org/themes/functionality/sidebars/#registering-a-sidebar
  */
 function cover2_widgets_init() {
-	// Overlay widgets.
+	// Overlay widget.
 	register_sidebar( array(
 		'name'          => esc_html__( 'Overlay', 'cover2' ),
 		'id'            => 'sidebar-overlay',
+		'description'   => __( 'Add widgets here to appear in your sidebar.', 'cover2' ),
+		'before_widget' => '<section id="%1$s" class="widget %2$s">',
+		'after_widget'  => '</section>',
+		'before_title'  => '<h2 class="widget-title">',
+		'after_title'   => '</h2>',
+	) );
+	
+	// Overlay widget.
+	register_sidebar( array(
+		'name'          => esc_html__( 'Footer (Left)', 'cover2' ),
+		'id'            => 'sidebar-footer-1',
+		'description'   => __( 'Add widgets here to appear in your sidebar.', 'cover2' ),
+		'before_widget' => '<section id="%1$s" class="widget %2$s">',
+		'after_widget'  => '</section>',
+		'before_title'  => '<h2 class="widget-title">',
+		'after_title'   => '</h2>',
+	) );
+	register_sidebar( array(
+		'name'          => esc_html__( 'Footer (Middle)', 'cover2' ),
+		'id'            => 'sidebar-footer-2',
+		'description'   => __( 'Add widgets here to appear in your sidebar.', 'cover2' ),
+		'before_widget' => '<section id="%1$s" class="widget %2$s">',
+		'after_widget'  => '</section>',
+		'before_title'  => '<h2 class="widget-title">',
+		'after_title'   => '</h2>',
+	) );
+	register_sidebar( array(
+		'name'          => esc_html__( 'Footer (Right)', 'cover2' ),
+		'id'            => 'sidebar-footer-3',
 		'description'   => __( 'Add widgets here to appear in your sidebar.', 'cover2' ),
 		'before_widget' => '<section id="%1$s" class="widget %2$s">',
 		'after_widget'  => '</section>',